Address log viewer review feedback
