+++ apps/favn_view/lib/favn_view/logs_live_support.ex
@@ -7,6 +7,7 @@ defmodule FavnView.LogsLiveSupport do
   alias Favn.Log.Filter
   alias FavnView.Components.AssetCataloguePage
   alias FavnView.LogsViewModel
+  alias FavnView.RunStepViewModel
 
   @initial_limit 200
   @fetch_limit 500
@@ -131,7 +132,7 @@ defmodule FavnView.LogsLiveSupport do
   end
 
   defp load_initial_logs(filter) do
-    case FavnOrchestrator.list_logs(filter, limit: @fetch_limit) do
+    case FavnOrchestrator.list_logs(filter, limit: @fetch_limit, order: :desc) do
       {:ok, %{items: items}} ->
         %{status: :ready, logs: LogsViewModel.trim_latest(items, @initial_limit)}
 
@@ -213,7 +214,7 @@ defmodule FavnView.LogsLiveSupport do
     finished_at = Map.get(run, :finished_at)
     status = Map.get(run, :status)
     target = target_label(Map.get(run, :asset_ref), Map.get(run, :target_refs, []))
-    asset_results = run_step_results(run, run.id)
+    asset_results = RunStepViewModel.from_run(run)
 
     %{
       found?: true,
@@ -227,94 +228,6 @@ defmodule FavnView.LogsLiveSupport do
       asset_results: asset_results
     }
   end
-
-  defp run_step_results(run, run_id) do
-    node_results = node_results(Map.get(run, :node_results, %{}), run_id)
-
-    if node_results == [] do
-      asset_results(Map.get(run, :asset_results, %{}), run_id)
-    else
-      node_results
-    end
-  end
-
-  defp node_results(results, run_id) when is_map(results) do
-    results
-    |> Enum.map(fn {key, result} -> node_result(result, run_id, key) end)
-    |> Enum.sort_by(& &1.asset_ref)
-  end
-
-  defp node_results(results, run_id) when is_list(results) do
-    results |> Enum.map(&node_result(&1, run_id, node_key(&1))) |> Enum.sort_by(& &1.asset_ref)
-  end
-
-  defp node_results(_results, _run_id), do: []
-
-  defp node_result(result, run_id, node_key) do
-    asset_ref = LogsViewModel.ref_label(Map.get(result, :ref) || node_asset_ref(node_key))
-
-    %{
-      id: step_id(result, run_id, node_key, asset_ref),
-      asset_ref: asset_ref,
-      display_name: LogsViewModel.display_name(asset_ref),
-      status: LogsViewModel.status_label(Map.get(result, :status)),
-      status_tone: LogsViewModel.status_tone(Map.get(result, :status)),
-      duration: LogsViewModel.duration_ms_label(Map.get(result, :duration_ms)),
-      started_at: LogsViewModel.timestamp_label(Map.get(result, :started_at)),
-      attempt: Map.get(result, :attempt)
-    }
-  end
-
-  defp asset_results(results, run_id) when is_map(results) do
-    results
-    |> Enum.map(fn {key, result} -> asset_result(result, run_id, key) end)
-    |> Enum.sort_by(& &1.asset_ref)
-  end
-
-  defp asset_results(results, run_id) when is_list(results) do
-    results |> Enum.map(&asset_result(&1, run_id, nil)) |> Enum.sort_by(& &1.asset_ref)
-  end
-
-  defp asset_results(_results, _run_id), do: []
-
-  defp asset_result(result, run_id, key) do
-    asset_ref = LogsViewModel.ref_label(Map.get(result, :ref) || key)
-
-    %{
-      id: step_id(result, run_id, key, asset_ref),
-      asset_ref: asset_ref,
-      display_name: LogsViewModel.display_name(asset_ref),
-      status: LogsViewModel.status_label(Map.get(result, :status)),
-      status_tone: LogsViewModel.status_tone(Map.get(result, :status)),
-      duration: LogsViewModel.duration_ms_label(Map.get(result, :duration_ms)),
-      started_at: LogsViewModel.timestamp_label(Map.get(result, :started_at)),
-      attempt: Map.get(result, :attempt)
-    }
-  end
-
-  defp step_id(result, run_id, key, asset_ref) do
-    Map.get(result, :id) || Map.get(result, "id") || Map.get(result, :step_id) ||
-      Map.get(result, "step_id") || persisted_key_id(key, asset_ref) ||
-      LogsViewModel.deterministic_step_id(run_id, asset_ref)
-  end
-
-  defp persisted_key_id(nil, _asset_ref), do: nil
-
-  defp persisted_key_id(key, asset_ref) when is_binary(key) and key != asset_ref,
-    do: LogsViewModel.safe_id(key)
-
-  defp persisted_key_id(key, _asset_ref) when is_tuple(key),
-    do:
-      key
-      |> :erlang.term_to_binary()
-      |> Base.url_encode64(padding: false)
-      |> LogsViewModel.safe_id()
-
-  defp persisted_key_id(_key, _asset_ref), do: nil
-
-  defp node_key(result), do: Map.get(result, :node_key) || Map.get(result, "node_key")
-  defp node_asset_ref({ref, _window}) when is_tuple(ref), do: ref
-  defp node_asset_ref(_node_key), do: nil
 
   defp asset_facts(nil), do: []
